update value types

+++ domain/src/main/scala/com/shmrkm/chatworkWebhook/domain/model/account/FromAccountId.scala
@@ -1,3 +1,3 @@
 package com.shmrkm.chatworkWebhook.domain.model.account
 
-case class FromAccountId(value: Int)
+case class FromAccountId(value: Long)
